+++ packages/assignment-5/src/refactoring/components/CartPage.tsx
@@ -1,6 +1,6 @@
 import { Coupon, Product } from '../../types.ts';
 import { useCart } from '../hooks/useCart.ts';
-import { getAppliedDiscount } from '../hooks/utils/cartUtils.ts';
+import CartItemComponent from './cart/CartItem.tsx';
 import ProductItem from './cart/ProductItem.tsx';
 
 interface Props {
@@ -27,6 +27,7 @@ export const CartPage = ({ products, coupons }: Props) => {
         <div className="space-y-2">
           {products.map((product) => (
             <ProductItem
+              key={product.id}
               product={product}
               cart={cart}
               addToCart={() => addToCart(product)}
@@ -38,52 +39,14 @@ export const CartPage = ({ products, coupons }: Props) => {
         <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
 
         <div className="space-y-2">
-          {cart.map((item) => {
-            const appliedDiscount = getAppliedDiscount(item);
-            return (
-              <div
-                key={item.product.id}
-                className="flex justify-between items-center bg-white p-3 rounded shadow"
-              >
-                <div>
-                  <span className="font-semibold">{item.product.name}</span>
-                  <br />
-                  <span className="text-sm text-gray-600">
-                    {item.product.price}원 x {item.quantity}
-                    {appliedDiscount > 0 && (
-                      <span className="text-green-600 ml-1">
-                        ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
-                      </span>
-                    )}
-                  </span>
-                </div>
-                <div>
-                  <button
-                    onClick={() =>
-                      updateQuantity(item.product.id, item.quantity - 1)
-                    }
-                    className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                  >
-                    -
-                  </button>
-                  <button
-                    onClick={() =>
-                      updateQuantity(item.product.id, item.quantity + 1)
-                    }
-                    className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                  >
-                    +
-                  </button>
-                  <button
-                    onClick={() => removeFromCart(item.product.id)}
-                    className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                  >
-                    삭제
-                  </button>
-                </div>
-              </div>
-            );
-          })}
+          {cart.map((item) => (
+            <CartItemComponent
+              key={item.product.id}
+              item={item}
+              updateQuantity={updateQuantity}
+              removeFromCart={removeFromCart}
+            />
+          ))}
         </div>
 
         <div className="mt-6 bg-white p-4 rounded shadow">
